crud - I am now routing from one application but I left a hook for plugins with MobileApplication.
